requirements.txt
# Note that GDAL is installed in Dockerfile to match the binary GDAL version
arrow==1.3.0
asgiref==3.8.1
blessed==1.20.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.7
django-environ==0.11.2
django-filter==24.3
django-leaflet==0.31.0
dj-database-url==2.3.0
django-picklefield==3.2
django-q2==1.7.4
django-sesame==3.2.2
djangorestframework==3.15.2
djangorestframework-gis==1.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1
django-currentuser==0.8.0
django-import-export==4.3.3
gunicorn==23.0.0
idna==3.10
pyasn1==0.6.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
types-python-dateutil==2.9.0.20241206
urllib3==2.3.0
wcwidth==0.2.13
whitenoise==6.8.2
psycopg2-binary==2.9.10
packaging==21.3
pycryptodome==3.21.0
pyalex>=0.4.0
python-stdnum>=2.0.0